# px fuse driver optimized for fastpath in blkmq mode