--- a/pxd.c
+++ b/pxd.c
@@ -382,7 +382,7 @@ static const struct block_device_operations pxd_bd_ops = {
 	.open			= pxd_open,
 	.release		= pxd_release,
 #if LINUX_VERSION_CODE >= KERNEL_VERSION(5,9,0) && defined(__PXD_BIO_MAKE_REQ__)
-	.submit_io		= pxd_bio_make_request_entryfn,
+	.submit_bio		= pxd_bio_make_request_entryfn,
 #endif
 };
 
@@ -601,7 +601,7 @@ int pxd_handle_device_limits(struct fuse_req *req, uint32_t *size, uint64_t *off
 		}
 
 		bio_chain(b, req->bio);
-#if LINUX_VERSION_CODE <= KERNEL_VERSION(5,9,0)
+#if LINUX_VERSION_CODE < KERNEL_VERSION(5,9,0)
 		generic_make_request(b);
 #else
 		submit_bio_noacct(b);
@@ -1309,8 +1309,8 @@ ssize_t pxd_add(struct fuse_conn *fc, struct pxd_add_ext_out *add)
 	pxd_dev->ctx = ctx;
 	pxd_dev->connected = true; // fuse slow path connection
 	pxd_dev->size = add->size;
-	pxd_dev->mode = add->open_mode;
 	pxd_dev->fastpath = add->enable_fp;
+	pxd_dev->mode = add->open_mode;
 
 	// congestion init
 	init_waitqueue_head(&pxd_dev->suspend_wq);
@@ -1470,8 +1470,12 @@ ssize_t pxd_ioc_update_size(struct fuse_conn *fc, struct pxd_update_size *update
 	set_capacity(pxd_dev->disk, update_size->size / SECTOR_SIZE);
 	spin_unlock(&pxd_dev->lock);
 
+#if LINUX_VERSION_CODE >= KERNEL_VERSION(5,10,0)
+	revalidate_disk_size(pxd_dev->disk, true);
+#else
 	err = revalidate_disk(pxd_dev->disk);
 	BUG_ON(err);
+#endif
 	put_device(&pxd_dev->dev);
 
 	return 0;