support fastpath in blkmq mode as well.

1/ code reorg - to take out common lib/helper code separate. (kiolib.[ch] files)
2/ keep pxd_fastpath common framework to handle IO in both makereq and blkmq mode.
3/ keep make request IO path that supports both native/fast in separate pxd_bio_makereq.[ch] file.
4/ develop blkmq IO path to support native/fast path all in separate pxd_bio_blkmq.[ch]
5/ remove pxd_fastpath_stub code since fastpath is supported always in all modes.
6/ default enable blkmq mode handling, allows to deprecate makereq mode.

https://portworx.atlassian.net/browse/PWX-18616

For master branch: ln/iopath-opt (full changes).
This has been split into 3 changes over master branch.
https://github.com/portworx/px-fuse/pull/203
https://github.com/portworx/px-fuse/pull/204
https://github.com/portworx/px-fuse/pull/205

# test runs # 
https://jenkins.portworx.dev/job/DEV/job/Porx-02/653 - dmthin bvt - for regression - PASS
https://jenkins.portworx.dev/job/DEV/job/Porx-01/900 - btrfs bvt - for regression - PASS
https://jenkins.portworx.dev/job/DEV/job/Porx-01-caching/464 - px-cache bvt - for regression - PASS


fastpath local (clone in fastpath regression is fixed with https://github.com/portworx/porx/pull/7542)
https://jenkins.portworx.dev/job/DEV/job/Porx-02/658 - dmthin bvt - PASSED
https://jenkins.portworx.dev/job/DEV/job/Porx-02/660 - dmthin bvt - PASSED
https://jenkins.portworx.dev/job/DEV/job/Porx-03/558 - btrfs bvt - PASSED
https://jenkins.portworx.dev/job/DEV/job/Porx-01-caching/469 - px-cache bvt - PASSED

fastpath nvmeof-tcp (allow multi replicas in fastpath true)
https://jenkins.portworx.dev/job/DEV/job/Porx-01/907 - dmthin bvt - PASSED

fastpath nvmeof-tcp(single replica)
https://jenkins.portworx.dev/job/DEV/job/Porx-01-caching/471/ - px-cache bvt - PASSED
https://jenkins.portworx.dev/job/DEV/job/Porx-03/562 - btrfs bvt - PASSED

TestCloneSticky seems a regression that needs to be fixed. (fixed with https://github.com/portworx/porx/pull/7542)
The IO path and driver is stable with fastpath enabled.

```
dmthin.
nvme baseline:
                   dmthin-pxvol(ext4)     md(ext4)            2.7-fastpath    2.7 nativepath    proposed-fastpath proposed-nativepath
seqwrite     406MiB/s                     863MiB/s           375MiB/s          391MiB/s           815MiB/s                  384MiB/s 
randwrite   350MiB/s                     689MiB/s           327MiB/s          372MiB/s           683MiB/s                   373MiB/s
seqread     453MiB/s                     943MiB/s           427MiB/s           273MiB/s           839MiB/s                   271MiB/s
randread    386MiB/s                    749MiB/s           364MiB/s           265MiB/s           675MiB/s                   269MiB/s 

** no regression in nativepath
** proposed fastpath is blkmq mode for both nativepath and fastpath
** over dmthin
```

```
[root@slave0 px-fuse]# uname -a
Linux slave0 4.20.13-1.el7.elrepo.x86_64 #1 SMP Wed Feb 27 10:02:05 EST 2019 x86_64 x86_64 x86_64 GNU/Linux
[root@slave0 px-fuse]#


ssd:
2.7.0     makereq(fastpath)   makereq(failover)   blkmq(nativepath)     
seqwrite: 359MiB/s            230MiB/s              234MiB/s 
randwrite:313MiB/s            324MiB/s              312MiB/s
seqread:  407MiB/s            226MiB/s              228MiB/s
randread: 317MiB/s            226MiB/s              217MiB/s

new consolidated driver:
blkmq mode:  fastpath   failover      blkmq-nativepath
seqwrite:    425MiB/s   375MiB/s        389MiB/s
randwrite:   317MiB/s   313MiB/s        316MiB/s
seqread:     475MiB/s   234MiB/s        235MiB/s
randread:    371MiB/s   214MiB/s        217MiB/s

makereq mode: fastpath   failover    makereq-nativepath
seqwrite:      426MiB/s   391MiB/s    403MiB/s 
randwrite:     317MiB/s   332MiB/s    337MiB/s
seqread:       309MiB/s   216MiB/s    213MiB/s
randread:      251MiB/s   202MiB/s    201MiB/s

blkmq or makereq combined mode for fastpath shows better performance on ssd over nativepath.


nvme:

2.7.0     makereq(fastpath)   makereq(failover)   blkmq(nativepath)     
seqwrite:    375MiB/s           390MiB/s            391MiB/s
randwrite:   327MiB/s           337MiB/s            372MiB/s
seqread:     427MiB/s           277MiB/s            273MiB/s
randread:    364MiB/s           270MiB/s            265MiB/s

new consolidated driver:
blkmq mode:  fastpath   failover      blkmq-nativepath
seqwrite:     815MiB/s  376MiB/s      384MiB/s 
randwrite:    683MiB/s  376MiB/s      373MiB/s
seqread:      839MiB/s  311MiB/s      271MiB/s
randread:     675MiB/s  299MiB/s      269MiB/s 


makereq mode: fastpath   failover    makereq-nativepath
seqwrite:      405MiB/s   396MiB/s    280MiB/s
randwrite:     333MiB/s   338MiB/s    334MiB/s
seqread:       415MiB/s   279MiB/s    277MiB/s
randread:      349MiB/s   269MiB/s    271MiB/s

blkmq combined mode for fastpath beats nativepath performance by many orders on nvme drives.
```

Tested kernels:
1/ Linux localhost.localdomain 3.10.0-1160.el7.x86_64 #1 SMP Mon Oct 19 16:18:59 UTC 2020 x86_64 x86_64 x86_64 GNU/Linux

Automation runs:
(in fastpath) https://jenkins.portworx.dev/job/DEV/job/Porx-01-caching/417 - PASSED - smoke test, centos 7.0 , 3.10 kernel
(fastpath) https://jenkins.portworx.dev/job/DEV/job/Porx-01-caching/418 - PASSED - smoke test, 4.20 kernel, btrfs
(fastpath) https://jenkins.portworx.dev/job/DEV/job/Porx-01-caching/419 - PASSED - smoke test, 4.20 kernel, dmthin


Test logs:
1a) 3.10 kernel in make request mode - px console
```
[root@localhost px-fuse]# insmod ./px.ko
[root@localhost px-fuse]# systemctl restart portworx
[root@localhost px-fuse]# pxctl v l
ID                      NAME    SIZE    HA      SHARED  ENCRYPTED       PROXY-VOLUME    IO_PRIORITY     STATUS          SNAP-ENABLED
25476599517741332       vol1    1 GiB   1       no      no              no              HIGH            up - detached   no
[root@localhost px-fuse]# pxctl v c vol2
Volume successfully created: 487238633571446003
[root@localhost px-fuse]# pxctl host attach vol2
Volume successfully attached at: /dev/pxd/pxd487238633571446003
[root@localhost px-fuse]# pxctl v i -e vol2
        Volume                   :  487238633571446003
        Name                     :  vol2
        Size                     :  1.0 GiB
        Format                   :  ext4
        HA                       :  1
        IO Priority              :  HIGH
        Creation time            :  Feb 11 11:11:25 UTC 2021
        Shared                   :  no
        Status                   :  up
        State                    :  Attached: 00255795-0c09-45ac-8880-53783a4df9c8 (10.0.3.15)
        Last Attached            :  Feb 11 11:11:34 UTC 2021
        Device Path              :  /dev/pxd/pxd487238633571446003
        Mount Options            :  discard
        ScanPolicy               :  none
        ProxyWrite               :  false
        Reads                    :  0
        Reads MS                 :  0
        Bytes Read               :  0
        Writes                   :  0
        Writes MS                :  0
        Bytes Written            :  0
        IOs in progress          :  0
        Bytes used               :  33 MiB
        Replica sets on nodes:
                Set 0
                  Node           : 10.0.3.15 (Pool 510d57bf-4dfd-40af-b901-c3f7fdca2793 )
        Replication Status       :  Up

        Displaying extended volume state:
        Fastpath preferred      : true
        Fastpath promoted       : true
        Fastpath dirty          : false
        Fastpath attached       : FASTPATH_ACTIVE
        Fastpath coordinator    : 00255795-0c09-45ac-8880-53783a4df9c8
        Fastpath replicas       : 1
        Fastpath replica property follows:
                Replica : 0
                        On Node         : 00255795-0c09-45ac-8880-53783a4df9c8
                        Protocol        : FASTPATH_PROTO_ISCSI
                        Secure          : true
                        Exported        : true
                                Target  : /var/.px/0/487238633571446003/pxdev
                                Source  : /var/.px/0/487238633571446003/pxdev
                                Type    : File
                        Imported        : true
                                Mapped local device: /var/.px/0/487238633571446003/pxdev
[root@localhost px-fuse]# mkdir -p /mnt/vol2; pxctl host mount --path=/mnt/vol2 vol2
Volume vol2 successfully mounted at /mnt/vol2
[root@localhost px-fuse]# fio --bs=4k --ioengine=libaio --iodepth=16 --rw=write --rwmixwrite=50 --size=800M --direct=1 --verify=md5 --do_verify=1 --direct=1 --directory=/mnt/vol2 --name=test
test: (g=0): rw=write, bs=(R) 4096B-4096B, (W) 4096B-4096B, (T) 4096B-4096B, ioengine=libaio, iodepth=16
fio-3.7
Starting 1 process
test: Laying out IO file (1 file / 800MiB)
Jobs: 1 (f=1): [V(1)][100.0%][r=458MiB/s,w=0KiB/s][r=117k,w=0 IOPS][eta 00m:00s]
test: (groupid=0, jobs=1): err= 0: pid=16759: Thu Feb 11 06:12:04 2021
   read: IOPS=117k, BW=458MiB/s (481MB/s)(800MiB/1745msec)
    slat (usec): min=2, max=1603, avg= 2.77, stdev= 6.07
    clat (usec): min=5, max=2289, avg=128.04, stdev=38.68
     lat (usec): min=7, max=2310, avg=130.87, stdev=39.57
    clat percentiles (usec):
     |  1.00th=[  111],  5.00th=[  112], 10.00th=[  112], 20.00th=[  112],
     | 30.00th=[  113], 40.00th=[  114], 50.00th=[  114], 60.00th=[  116],
     | 70.00th=[  121], 80.00th=[  135], 90.00th=[  172], 95.00th=[  196],
     | 99.00th=[  239], 99.50th=[  260], 99.90th=[  482], 99.95th=[  701],
     | 99.99th=[ 1074]
  write: IOPS=62.4k, BW=244MiB/s (256MB/s)(800MiB/3283msec)
    slat (usec): min=10, max=3075, avg=13.03, stdev=22.14
    clat (usec): min=10, max=11934, avg=242.90, stdev=417.52
     lat (usec): min=42, max=11948, avg=256.00, stdev=418.53
    clat percentiles (usec):
     |  1.00th=[  188],  5.00th=[  190], 10.00th=[  190], 20.00th=[  190],
     | 30.00th=[  192], 40.00th=[  194], 50.00th=[  200], 60.00th=[  212],
     | 70.00th=[  227], 80.00th=[  251], 90.00th=[  281], 95.00th=[  314],
     | 99.00th=[  474], 99.50th=[  955], 99.90th=[10683], 99.95th=[10945],
     | 99.99th=[11207]
   bw (  KiB/s): min=132988, max=276992, per=93.66%, avg=233697.86, stdev=49631.60, samples=7
   iops        : min=33247, max=69248, avg=58424.43, stdev=12407.89, samples=7
  lat (usec)   : 10=0.01%, 20=0.01%, 50=0.01%, 100=0.01%, 250=89.44%
  lat (usec)   : 500=10.05%, 750=0.18%, 1000=0.07%
  lat (msec)   : 2=0.09%, 4=0.08%, 10=0.02%, 20=0.06%
  cpu          : usr=49.57%, sys=22.38%, ctx=422377, majf=0, minf=4840
  IO depths    : 1=0.1%, 2=0.1%, 4=0.1%, 8=0.1%, 16=100.0%, 32=0.0%, >=64=0.0%
     submit    : 0=0.0%, 4=100.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=0.0%
     complete  : 0=0.0%, 4=100.0%, 8=0.0%, 16=0.1%, 32=0.0%, 64=0.0%, >=64=0.0%
     issued rwts: total=204800,204800,0,0 short=0,0,0,0 dropped=0,0,0,0
     latency   : target=0, window=0, percentile=100.00%, depth=16

Run status group 0 (all jobs):
   READ: bw=458MiB/s (481MB/s), 458MiB/s-458MiB/s (481MB/s-481MB/s), io=800MiB (839MB), run=1745-1745msec
  WRITE: bw=244MiB/s (256MB/s), 244MiB/s-244MiB/s (256MB/s-256MB/s), io=800MiB (839MB), run=3283-3283msec

Disk stats (read/write):
  pxd!pxd487238633571446003: ios=190026/204800, merge=0/0, ticks=229/632, in_queue=858, util=16.95%
[root@localhost px-fuse]# pxctl host attach --fastpath=false vol2
Volume successfully attached at: /dev/pxd/pxd487238633571446003
[root@localhost px-fuse]# pxctl v i -e vol2
        Volume                   :  487238633571446003
        Name                     :  vol2
        Size                     :  1.0 GiB
        Format                   :  ext4
        HA                       :  1
        IO Priority              :  HIGH
        Creation time            :  Feb 11 11:11:25 UTC 2021
        Shared                   :  no
        Status                   :  up
        State                    :  Attached: 00255795-0c09-45ac-8880-53783a4df9c8 (10.0.3.15)
        Last Attached            :  Feb 11 11:12:19 UTC 2021
        Device Path              :  /dev/pxd/pxd487238633571446003
        Mount Options            :  discard
        ScanPolicy               :  none
        ProxyWrite               :  false
        Reads                    :  204856
        Reads MS                 :  248
        Bytes Read               :  839974912
        Writes                   :  204874
        Writes MS                :  838
        Bytes Written            :  855814144
        IOs in progress          :  0
        Bytes used               :  849 MiB
        Replica sets on nodes:
                Set 0
                  Node           : 10.0.3.15 (Pool 510d57bf-4dfd-40af-b901-c3f7fdca2793 )
        Replication Status       :  Up

        Displaying extended volume state:
        Fastpath preferred      : true
        Fastpath promoted       : false
        Fastpath dirty          : true
        Fastpath attached       : FASTPATH_INACTIVE
[root@localhost px-fuse]# fio --bs=4k --ioengine=libaio --iodepth=16 --rw=write --rwmixwrite=50 --size=800M --direct=1 --verify=md5 --do_verify=1 --direct=1 --directory=/mnt/vol2 --name=test
test: (g=0): rw=write, bs=(R) 4096B-4096B, (W) 4096B-4096B, (T) 4096B-4096B, ioengine=libaio, iodepth=16
fio-3.7
Starting 1 process
Jobs: 1 (f=1): [V(1)][100.0%][r=217MiB/s,w=0KiB/s][r=55.4k,w=0 IOPS][eta 00m:00s]
test: (groupid=0, jobs=1): err= 0: pid=16836: Thu Feb 11 06:12:49 2021
   read: IOPS=55.2k, BW=216MiB/s (226MB/s)(800MiB/3709msec)
    slat (nsec): min=1075, max=419578, avg=1809.08, stdev=2245.29
    clat (usec): min=113, max=19713, avg=282.20, stdev=123.19
     lat (usec): min=115, max=19719, avg=284.06, stdev=123.60
    clat percentiles (usec):
     |  1.00th=[  221],  5.00th=[  223], 10.00th=[  225], 20.00th=[  229],
     | 30.00th=[  241], 40.00th=[  269], 50.00th=[  273], 60.00th=[  281],
     | 70.00th=[  281], 80.00th=[  293], 90.00th=[  314], 95.00th=[  383],
     | 99.00th=[  644], 99.50th=[  676], 99.90th=[  873], 99.95th=[ 1172],
     | 99.99th=[ 2933]
  write: IOPS=45.6k, BW=178MiB/s (187MB/s)(800MiB/4495msec)
    slat (usec): min=6, max=1518, avg= 8.23, stdev= 9.12
    clat (usec): min=143, max=17629, avg=342.53, stdev=535.19
     lat (usec): min=150, max=17637, avg=350.82, stdev=535.46
    clat percentiles (usec):
     |  1.00th=[  255],  5.00th=[  260], 10.00th=[  260], 20.00th=[  265],
     | 30.00th=[  269], 40.00th=[  281], 50.00th=[  289], 60.00th=[  306],
     | 70.00th=[  326], 80.00th=[  347], 90.00th=[  383], 95.00th=[  429],
     | 99.00th=[  734], 99.50th=[ 1434], 99.90th=[10945], 99.95th=[11600],
     | 99.99th=[13173]
   bw (  KiB/s): min=174730, max=188790, per=99.78%, avg=181843.22, stdev=5341.33, samples=9
   iops        : min=43682, max=47197, avg=45460.56, stdev=1335.41, samples=9
  lat (usec)   : 250=17.59%, 500=79.32%, 750=2.50%, 1000=0.18%
  lat (msec)   : 2=0.21%, 4=0.04%, 10=0.08%, 20=0.08%
  cpu          : usr=31.35%, sys=10.96%, ctx=26153, majf=0, minf=4842
  IO depths    : 1=0.1%, 2=0.1%, 4=0.1%, 8=0.1%, 16=100.0%, 32=0.0%, >=64=0.0%
     submit    : 0=0.0%, 4=100.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=0.0%
     complete  : 0=0.0%, 4=100.0%, 8=0.0%, 16=0.1%, 32=0.0%, 64=0.0%, >=64=0.0%
     issued rwts: total=204800,204800,0,0 short=0,0,0,0 dropped=0,0,0,0
     latency   : target=0, window=0, percentile=100.00%, depth=16

Run status group 0 (all jobs):
   READ: bw=216MiB/s (226MB/s), 216MiB/s-216MiB/s (226MB/s-226MB/s), io=800MiB (839MB), run=3709-3709msec
  WRITE: bw=178MiB/s (187MB/s), 178MiB/s-178MiB/s (187MB/s-187MB/s), io=800MiB (839MB), run=4495-4495msec

Disk stats (read/write):
  pxd!pxd487238633571446003: ios=201440/204803, merge=0/0, ticks=0/0, in_queue=0, util=0.00%
[root@localhost px-fuse]# pxctl host attach vol2
Volume successfully attached at: /dev/pxd/pxd487238633571446003
[root@localhost px-fuse]# pxctl v i -e vol2
        Volume                   :  487238633571446003
        Name                     :  vol2
        Size                     :  1.0 GiB
        Format                   :  ext4
        HA                       :  1
        IO Priority              :  HIGH
        Creation time            :  Feb 11 11:11:25 UTC 2021
        Shared                   :  no
        Status                   :  up
        State                    :  Attached: 00255795-0c09-45ac-8880-53783a4df9c8 (10.0.3.15)
        Last Attached            :  Feb 11 11:13:03 UTC 2021
        Device Path              :  /dev/pxd/pxd487238633571446003
        Mount Options            :  discard
        ScanPolicy               :  none
        ProxyWrite               :  false
        Reads                    :  409656
        Reads MS                 :  248
        Bytes Read               :  1678835712
        Writes                   :  409677
        Writes MS                :  838
        Bytes Written            :  1694687232
        IOs in progress          :  0
        Bytes used               :  849 MiB
        Replica sets on nodes:
                Set 0
                  Node           : 10.0.3.15 (Pool 510d57bf-4dfd-40af-b901-c3f7fdca2793 )
        Replication Status       :  Up

        Displaying extended volume state:
        Fastpath preferred      : true
        Fastpath promoted       : true
        Fastpath dirty          : false
        Fastpath attached       : FASTPATH_ACTIVE
        Fastpath coordinator    : 00255795-0c09-45ac-8880-53783a4df9c8
        Fastpath replicas       : 1
        Fastpath replica property follows:
                Replica : 0
                        On Node         : 00255795-0c09-45ac-8880-53783a4df9c8
                        Protocol        : FASTPATH_PROTO_ISCSI
                        Secure          : true
                        Exported        : true
                                Target  : /var/.px/0/487238633571446003/pxdev
                                Source  : /var/.px/0/487238633571446003/pxdev
                                Type    : File
                        Imported        : true
                                Mapped local device: /var/.px/0/487238633571446003/pxdev
[root@localhost px-fuse]# fio --bs=4k --ioengine=libaio --iodepth=16 --rw=write --rwmixwrite=50 --size=800M --direct=1 --verify=md5 --do_verify=1 --direct=1 --directory=/mnt/vol2 --name=test
test: (g=0): rw=write, bs=(R) 4096B-4096B, (W) 4096B-4096B, (T) 4096B-4096B, ioengine=libaio, iodepth=16
fio-3.7
Starting 1 process
Jobs: 1 (f=1): [V(1)][100.0%][r=458MiB/s,w=0KiB/s][r=117k,w=0 IOPS][eta 00m:00s]
test: (groupid=0, jobs=1): err= 0: pid=16892: Thu Feb 11 06:13:13 2021
   read: IOPS=117k, BW=458MiB/s (480MB/s)(800MiB/1747msec)
    slat (usec): min=2, max=1703, avg= 2.74, stdev= 6.10
    clat (usec): min=5, max=1833, avg=128.16, stdev=37.71
     lat (usec): min=7, max=1836, avg=130.97, stdev=38.50
    clat percentiles (usec):
     |  1.00th=[  111],  5.00th=[  111], 10.00th=[  111], 20.00th=[  112],
     | 30.00th=[  112], 40.00th=[  113], 50.00th=[  114], 60.00th=[  116],
     | 70.00th=[  124], 80.00th=[  141], 90.00th=[  174], 95.00th=[  196],
     | 99.00th=[  237], 99.50th=[  258], 99.90th=[  441], 99.95th=[  515],
     | 99.99th=[ 1221]
  write: IOPS=80.1k, BW=313MiB/s (328MB/s)(800MiB/2558msec)
    slat (usec): min=8, max=12813, avg=11.77, stdev=105.37
    clat (nsec): min=832, max=13396k, avg=187611.53, stdev=419715.15
     lat (usec): min=11, max=13408, avg=199.46, stdev=433.76
    clat percentiles (usec):
     |  1.00th=[  141],  5.00th=[  141], 10.00th=[  141], 20.00th=[  143],
     | 30.00th=[  143], 40.00th=[  145], 50.00th=[  147], 60.00th=[  159],
…