@sulakshm 2021-03-13
This is a new changes to get fastpath also work in blkmq mode.