@sulakshm 2021-05-07
changes went through 
https://github.com/portworx/px-fuse/pull/203
https://github.com/portworx/px-fuse/pull/204
https://github.com/portworx/px-fuse/pull/205

hence closing this.